chore: moved everything out of folders
