--- Data/fileInput3.mdb
@@ -1,27 +1,0 @@
-CREATE fileOutput3.txt;
-DATABASES;
-
-CREATE TABLE employee ( employee_id INT, employee_name TEXT, employee_city TEXT, employee_department TEXT, employee_salary INT);
-
-INSERT INTO employee(employee_id, employee_name, employee_city, employee_department, employee_salary) VALUES  (1, 'Jax', 'Kuala Lumpur', 'HR', 60000); 
-INSERT INTO employee(employee_id, employee_name, employee_city, employee_department, employee_salary) VALUES  (2, 'Calvin', 'Johor Bahru', 'IT', 75000) ;
-INSERT INTO (employee_id, employee_name, employee_city, employee_department, employee_salary) VALUES (3, 'Kee', 'George Town', 'Finance', 70000);
-INSERT INTO (employee_id, employee_name, employee_city, employee_department, employee_salary) VALUES (4, 'Jason', 'Kota Kinabalu', 'IT', 80000);
-
-SELECT * FROM employee;
-
-TABLES;
-
-UPDATE employee SET employee_salary = 72000 WHERE employee_id = 3;
-
-SELECT * FROM employee WHERE employee_department = 'IT';
-
-DELETE FROM employee WHERE employee_id = 2;
-
-SELECT * FROM employee;
-
-SELECT COUNT(*) AS total_employees FROM employee;
-
-INSERT INTO employee(employee_id, employee_name, employee_city, employee_department, employee_salary)  VALUES (5, 'Ali', 'Shah Alam', 'HR', 65000);
-
-SELECT * FROM employee;
