--- Data/fileOutput1.txt
@@ -1,18 +1,0 @@
-> CREATE fileOutput1.txt; 
-> DATABASES;
-C:\Users\User\VS Code Stuff\C++\Light-Mariadb-Interpreter\src\fileInput1.mdb
-
-> CREATE TABLE customer( customer_id INT, customer_name TEXT, customer_city TEXT, customer_state TEXT, customer_country TEXT, customer_phone TEXT, customer_email TEXT); 
-> TABLES;
-customer
-
-> INSERT INTO customer(customer_id,customer_name,customer_city,customer_state,customer_country,custom er_phone,customer_email) VALUES (1,'name1','city1','state1','country1','phone1','email1'); 
-> INSERT INTO customer(customer_id,customer_name,customer_city,customer_state,customer_country,custom er_phone,customer_email) VALUES (2,'name2','city2','state2','country2','phone2','email2'); 
-> INSERT INTO customer(customer_id,customer_name,customer_city,customer_state,customer_country,custom er_phone,customer_email) VALUES (3,'name3','city3','state3','country3','phone3','email3'); 
-> INSERT INTO customer(customer_id,customer_name,customer_city,customer_state,customer_country,custom er_phone,customer_email) VALUES ();
-
-> SELECT * FROM customer;
-customer_id, customer_name, customer_city, customer_state, customer_country, customer_phone, customer_email
-1, name1, city1, state1, country1, phone1, email1
-2, name2, city2, state2, country2, phone2, email2
-3, name3, city3, state3, country3, phone3, email3
