+++ fileInput2.mdb
@@ -1,0 +1,20 @@
+CREATE fileOutput2.txt; 
+DATABASES;
+
+CREATE TABLE customer( customer_id INT, customer_name TEXT, customer_city TEXT, customer_state TEXT, customer_country TEXT, customer_phone TEXT, customer_email TEXT);
+
+INSERT INTO customer(customer_id,customer_name,customer_city,customer_state,customer_country,custom er_phone,customer_email) VALUES (1,'name1','city1','state1','country1','phone1','email1'); 
+INSERT INTO customer(customer_id,customer_name,customer_city,customer_state,customer_country,custom er_phone,customer_email) VALUES (2,'name2','city2','state2','country2','phone2','email2'); 
+INSERT INTO customer(customer_id,customer_name,customer_city,customer_state,customer_country,customer_phone,customer_email) VALUES (3,'name3','city3','state3','country3','phone3','email3'); 
+INSERT INTO customer(customer_id,customer_name,customer_city,customer_state,customer_country,custom er_phone,customer_email) VALUES (4,'name4','city4','state4','country4','phone4','email4'); SELECT * FROM customer;
+
+TABLES;
+
+UPDATE customer SET customer_email='email333' WHERE customer_id=3; 
+SELECT * FROM customer;
+
+DELETE FROM customer WHERE customer_id=4; 
+SELECT * FROM customer;
+
+SELECT COUNT(*) FROM customer;
+
