+++ fileOutput3.txt
@@ -1,0 +1,48 @@
+> CREATE fileOutput3.txt;
+> DATABASES;
+C:\Users\User\VS Code Stuff\C++\Light-Mariadb-Interpreter\src\fileInput3.mdb
+
+> CREATE TABLE employee ( employee_id INT, employee_name TEXT, employee_city TEXT, employee_department TEXT, employee_salary INT);
+
+> INSERT INTO employee(employee_id, employee_name, employee_city, employee_department, employee_salary) VALUES  (1, 'Jax', 'Kuala Lumpur', 'HR', 60000); 
+> INSERT INTO employee(employee_id, employee_name, employee_city, employee_department, employee_salary) VALUES  (2, 'Calvin', 'Johor Bahru', 'IT', 75000) ;
+> INSERT INTO (employee_id, employee_name, employee_city, employee_department, employee_salary) VALUES (3, 'Kee', 'George Town', 'Finance', 70000);
+> INSERT INTO (employee_id, employee_name, employee_city, employee_department, employee_salary) VALUES (4, 'Jason', 'Kota Kinabalu', 'IT', 80000);
+
+> SELECT * FROM employee;
+employee_id, employee_name, employee_city, employee_department, employee_salary
+1,  'Jax',  'Kuala Lumpur',  'HR',  60000
+2,  'Calvin',  'Johor Bahru',  'IT',  75000
+3,  'Kee',  'George Town',  'Finance',  70000
+4,  'Jason',  'Kota Kinabalu',  'IT',  80000
+
+> TABLES;
+employee
+
+> UPDATE employee SET employee_salary = 72000 WHERE employee_id = 3;
+
+> SELECT * FROM employee WHERE employee_department = 'IT';
+employee_id, employee_name, employee_city, employee_department, employee_salary
+1,  'Jax',  'Kuala Lumpur',  'HR',  60000
+2,  'Calvin',  'Johor Bahru',  'IT',  75000
+3,  'Kee',  'George Town',  'Finance',  70000
+4,  'Jason',  'Kota Kinabalu',  'IT',  80000
+
+> DELETE FROM employee WHERE employee_id = 2;
+
+> SELECT * FROM employee;
+employee_id, employee_name, employee_city, employee_department, employee_salary
+1,  'Jax',  'Kuala Lumpur',  'HR',  60000
+3,  'Kee',  'George Town',  'Finance',  70000
+4,  'Jason',  'Kota Kinabalu',  'IT',  80000
+
+> SELECT COUNT(*) AS total_employees FROM employee;
+
+> INSERT INTO employee(employee_id, employee_name, employee_city, employee_department, employee_salary)  VALUES (5, 'Ali', 'Shah Alam', 'HR', 65000);
+
+> SELECT * FROM employee;
+employee_id, employee_name, employee_city, employee_department, employee_salary
+1,  'Jax',  'Kuala Lumpur',  'HR',  60000
+3,  'Kee',  'George Town',  'Finance',  70000
+4,  'Jason',  'Kota Kinabalu',  'IT',  80000
+5,  'Ali',  'Shah Alam',  'HR',  65000
